separate out study results from wrapper componneny results panel
