+++ package.json
@@ -1,6 +1,8 @@
 {
   "scripts": {
-    "test": "jest --coverage"
+    "test": "jest --coverage",
+    "lint": "eslint.js",
+    "fix-styles": "prettier-stylelint --write 'src/**/*.{css,scss}' "
   },
   "dependencies": {
     "@babel/preset-react": "^7.8.3",
@@ -12,13 +14,12 @@
     "@reach/router": "^1.3.3",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "camelcase-keys": "^6.1.2",
-    "core-js": "3",
+    "core-js": "^3.6.4",
     "ideogram": "1.11.0",
     "igv": "2.1.0",
     "jquery": "3.4.0",
     "jquery-ui": "1.12.1",
     "morpheus-app": "1.0.18",
-    "node-sass": "^4.13.1",
     "pluralize": "^8.0.0",
     "postcss-cssnext": "3.1.0",
     "prop-types": "^15.7.2",
@@ -28,16 +29,32 @@
     "react-clipboard.js": "^2.0.16",
     "react-compound-slider": "^3.0.0-beta.1",
     "react-dom": "^16.12.0",
+    "react-scripts": "^3.4.0",
     "react-table": "^7.0.0-rc.16",
-    "spin.js": "^4.0.0"
+    "spin.js": "^4.0.0",
+    "stylelint-config-sass-guidelines": "^7.0.0"
   },
   "devDependencies": {
+    "babel-eslint": "^10.0.3",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
+    "eslint": "^6.8.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-jsx-a11y": "*",
+    "eslint-plugin-prettier": "^3.1.2",
+    "eslint-plugin-react": "^7.18.3",
     "jest-cli": "^25.1.0",
     "jest-environment-enzyme": "^7.1.2",
     "jest-enzyme": "^7.1.2",
     "node-fetch": "^2.6.0",
-    "webpack-dev-server": "3.10.3"
+    "webpack-dev-server": "3.10.3",
+    "node-sass": "^4.13.1",
+    "prettier": "^1.19.1",
+    "prettier-eslint": "^9.0.1",
+    "prettier-stylelint": "^0.4.2",
+    "prop-types": "^15.7.2",
+    "stylelint-prettier": "^1.1.2"
   }
 }
